requirements.txt
google-genai>=1.0.0
requests
pymysql
ttkbootstrap
